pyproject.toml
[project]
name = "lixun-robot"
version = "0.1.0"
description = "基于知识库的智能问答机器人"
dependencies = [
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.2",
    "langchain-text-splitters>=0.0.1",
    "langgraph>=0.0.40",
    "faiss-cpu>=1.7.4",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "dashscope>=1.14.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
lixun-robot = "run:main"